Allowed setting logging configs through Docker
The 'ENV' Docker build argument has been renamed to 'BUILD_ENV'; it is separate from NODE_ENV, which is an environment variable that can be set apart from 'BUILD_ENV' if necessary.

There are now two new environment variables, LOG_LEVEL and LOG_FILTER, that override the settings in the NPM package config. This allows one to dynamically change the logging configuration from the environment instead of having to edit the script, which is especially useful for Docker, where it's much more convenient to edit the environment configuration in 'docker-compose.yml' or to pass them through 'docker run -e ...' than to edit 'package.json'.

The README has also been updated to reflect the latest configuration options.