Updated package dependencies, and updated .dockerignore
The .dockerignore now includes all files not essential to running the
Express server, for a slightly smaller package size. (Oddly enough,
however, the build for this version is larger than previous builds—not
sure why that is, although I've checked the internals after build and
they definitely include less files; most likely, an Alpine dependency
was updated in the meantime).

Additionally, changed name in 'package.json' to match README, and
stopped Alpine from running `apk upgrade` on build in the Dockerfile
to improve reliability.