Changed the way Docker builds are configured
Previously, the Docker image construction process was highly dependent
on 'package.json'; you had to edit the package every time you wanted to
change your build environment, and that didn’t play particularly well
with version control.

Now, the 'dk-build' script has been separated into 'dk-build-prod' and
'dk-build-dev', which build images with different environments; they
read the image version number from 'package.json', which *should* be
version-controlled (as opposed to the last environment the developer
built an image for). The resulting environment variables are saved in
'.env', which is untracked; one can then edit the '.env' to change
the environment or tag that the rest of the 'yarn dk-…' scripts use.